+++ src/main/kotlin/com/ticketera/dto/ticketTypes/UpdateTicketTypeDto.kt
@@ -2,13 +2,12 @@ package com.ticketera.dto.ticketTypes
 
 import com.ticketera.model.Event
 import com.ticketera.model.TicketType
-import java.math.BigDecimal
 import java.util.UUID
 
 data class UpdateTicketTypeDto(
     val id: UUID,
     val name: String?,
-    val price: BigDecimal?,
+    val price: Double?,
     val currency: String?,
     val startTime: Long?,
     val endTime: Long?,
